+++ TooManyAnimals/TooManyAnimals.cs
@@ -2,7 +2,6 @@ namespace StardewMods.TooManyAnimals;
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.Linq;
 using HarmonyLib;
 using StardewModdingAPI;