+++ BetterChests/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.1.0 (Unreleased)
 
+* Added Chest Finder.
 * Added localized tags for search and categorization.
 * Configure chest now happens from the chest inventory menu.
 * Fixed Shipping Bin not detected when launched by Chests Anywhere.